fix #182 enhance delay_response_milliseconds

+++ src/core/server/routing/rule_set/default_respond.rs
@@ -2,7 +2,7 @@ use serde::Deserialize;
 
 #[derive(Clone, Deserialize, Debug)]
 pub struct DefaultRespond {
-    pub delay_response_milliseconds: Option<u16>,
+    pub delay_response_milliseconds: Option<u32>,
 }
 
 impl DefaultRespond {
